Fixed ThemeSwitch somethimes displaying the wrong value
If prefers-color-scheme is supported is on the user's computer, ThemeProvider would make that the active theme. 
However, ThemeSwitcher did not account for this, leading it to sometimes displaying the wrong value before the user selected a theme.

Another way to fix this would be to set `enableSystem={false}` on ThemeProvider. This will disable system preference detection.